chore: 인터페이스 파일 추가
