+++ apps/ceos/src/pages/recruit/index.tsx
@@ -20,33 +20,14 @@ import { QueryClient, dehydrate, useQuery } from '@tanstack/react-query';
 import { recruitApi } from 'packages/utils';
 import Footer from '@ceos/components/Footer';
 import { useState } from 'react';
-import { PassDataInterface } from '@ceos/components/recruit/interface';
+import {
+  DateProps,
+  PassDataInterface,
+  RecruitStudyResponse,
+} from '@ceos/components/recruit/interface';
 import NonPass from '@ceos/components/recruit/nonpass';
 import DocPass from '@ceos/components/recruit/docpass';
 import FinPass from '@ceos/components/recruit/finpass';
-
-export interface RecruitStudyResponse {
-  generation: number;
-  prodStudyUrl: string;
-  designStudyUrl: string;
-  devStudyUrl: string;
-  startDateDoc: string;
-  endDateDoc: string;
-  resultDateDoc: string;
-  startDateInterview: string;
-  endDateInterview: string;
-  resultDateFinal: string;
-  openChatUrl: string;
-  otDate: string;
-  demoDayDate: string;
-}
-
-export interface DateProps {
-  startDateDoc: Date;
-  endDateDoc: Date;
-  resultDateDoc: Date;
-  resultDateFinal: Date;
-}
 
 const Recruit = () => {
   const { data, isLoading, isSuccess } = useQuery<RecruitStudyResponse>(
